+++ fastecdsa/encoding/asn1.py
@@ -106,8 +106,8 @@ def parse_asn1_length(data: bytes) -> Tuple[int, bytes, bytes]:
 
     if length > len(data):
         raise ASN1EncodingError(
-            "Parsed length of ASN.1 structure to be {} bytes but only {} bytes"
-            "remain in the provided data".format(length, len(data))
+            f"Parsed length of ASN.1 structure to be {length} bytes but only {len(data)} bytes"
+            f"remain in the provided data"
         )
 
     return length, data[:length], data[length:]
@@ -137,9 +137,7 @@ def parse_asn1_int(data: bytes) -> Tuple[int, bytes, bytes]:
     # integer length should match length indicated
     if length != len(data):
         raise ASN1EncodingError(
-            "Expected ASN.1 INTEGER to be {} bytes, got {} bytes".format(
-                length, len(data)
-            )
+            f"Expected ASN.1 INTEGER to be {length} bytes, got {len(data)} bytes"
         )
 
     return length, data, remaining