+++ fastecdsa/encoding/pem.py
@@ -1,6 +1,6 @@
 from binascii import a2b_base64, b2a_base64, hexlify
 from textwrap import wrap
-from typing import List, Optional, Tuple
+from typing import List, Tuple
 
 from . import KeyEncoder
 from .asn1 import (
@@ -21,12 +21,6 @@ from .asn1 import (
 from ..curve import Curve
 from ..point import Point
 
-EC_PRIVATE_HEADER = "-----BEGIN EC PRIVATE KEY-----"
-EC_PRIVATE_FOOTER = "-----END EC PRIVATE KEY-----"
-
-EC_PUBLIC_HEADER = "-----BEGIN PUBLIC KEY-----"
-EC_PUBLIC_FOOTER = "-----END PUBLIC KEY-----"
-
 
 class PEMEncoderError(Exception):
     def __init__(self, msg):
@@ -34,45 +28,54 @@ class PEMEncoderError(Exception):
 
 
 class PEMEncoder(KeyEncoder):
-    ASN1_PARSED_DATA: List[Tuple[bytes, bytes]] = []
-    binary_data = False
+    EC_PRIVATE_HEADER = b"-----BEGIN EC PRIVATE KEY-----"
+    EC_PRIVATE_FOOTER = b"-----END EC PRIVATE KEY-----"
 
-    @staticmethod
-    def _parse_ascii_armored_base64(data: str) -> bytes:
+    EC_PUBLIC_HEADER = b"-----BEGIN PUBLIC KEY-----"
+    EC_PUBLIC_FOOTER = b"-----END PUBLIC KEY-----"
+
+    def __init__(self):
+        self.asn1_parsed_data: List[Tuple[bytes, bytes]] = []
+
+    @classmethod
+    def _parse_ascii_armored_base64(cls, data: bytes) -> bytes:
         """Convert an ASCII armored key to raw binary data"""
         data = data.strip()
-        lines = (line for line in data.split("\n"))
+        lines = (line for line in data.split(b"\n"))
         next(lines).rstrip()  # header lines
 
-        base64_data = ""
+        base64_data = b""
         line = next(lines).rstrip()
 
-        while line and (line != EC_PRIVATE_FOOTER) and (line != EC_PUBLIC_FOOTER):
+        while (
+            line and (line != cls.EC_PRIVATE_FOOTER) and (line != cls.EC_PUBLIC_FOOTER)
+        ):
             base64_data += line
             line = next(lines).rstrip()
 
         return a2b_base64(base64_data)
 
-    @staticmethod
-    def _parse_asn1_structure(data: bytes):
+    def _parse_asn1_structure(self, data: bytes):
         """Recursively parse ASN.1 data"""
         data_type = data[:1]
         _, data, remaining = parse_asn1_length(data[1:])
 
         if data_type in [OCTET_STRING, BIT_STRING, OBJECT_IDENTIFIER]:
-            PEMEncoder.ASN1_PARSED_DATA.append((data_type, data))
+            self.asn1_parsed_data.append((data_type, data))
         elif data_type in [SEQUENCE, PUBLIC_KEY, PARAMETERS]:
-            PEMEncoder._parse_asn1_structure(data)
+            self._parse_asn1_structure(data)
 
         if remaining:
-            PEMEncoder._parse_asn1_structure(remaining)
+            self._parse_asn1_structure(remaining)
 
-    @staticmethod
-    def encode_public_key(Q: Point) -> str:
+    def encode_public_key(self, Q: Point) -> bytes:
         """Encode an EC public key as described in `RFC 5480 <https://tools.ietf.org/html/rfc5480>`_.
 
+        Args:
+            | Q (fastecdsa.point.Point): An ECDSA public key.
+
         Returns:
-            str: The ASCII armored encoded EC public key.
+            bytes: The ASCII armored encoded EC public key.
         """
         algorithm = asn1_ecpublickey()
         oid = asn1_oid(Q.curve)
@@ -83,30 +86,25 @@ class PEMEncoder(KeyEncoder):
         ec_public_key = asn1_structure(SEQUENCE, sequence)
         b64_data = "\n".join(wrap(b2a_base64(ec_public_key).decode(), 64))
 
-        return EC_PUBLIC_HEADER + "\n" + b64_data + "\n" + EC_PUBLIC_FOOTER
+        return (
+            self.EC_PUBLIC_HEADER
+            + b"\n"
+            + b64_data.encode()
+            + b"\n"
+            + self.EC_PUBLIC_FOOTER
+        )
 
-    @staticmethod
-    def encode_private_key(
-        d: int, Q: Optional[Point] = None, curve: Optional[Curve] = None
-    ) -> str:
-        """Encode an EC keypair as described in `RFC 5915 <https://tools.ietf.org/html/rfc5915.html>`_.
+    def encode_private_key(self, d: int, curve: Curve) -> bytes:
+        """Encode a private EC key as described in `RFC 5915 <https://tools.ietf.org/html/rfc5915.html>`_.
 
         Args:
-            | d (long): An ECDSA private key.
-            | Q (fastecdsa.point.Point): The ECDSA public key.
+            | d (int): An ECDSA private key.
             | curve (fastecdsa.curve.Curve): The curve that the private key is for.
 
         Returns:
-            str: The ASCII armored encoded EC keypair.
+            bytes: The ASCII armored encoded EC key.
         """
-        if Q is None and curve is None:
-            raise ValueError(
-                "Ambiguous encoding, public key or curve must passed as an argument"
-            )
-        elif Q is None:
-            Q = d * curve.G  # type: ignore
-        elif curve is None:
-            curve = Q.curve
+        Q: Point = d * curve.G
 
         version = asn1_ecversion()
         private_key = asn1_private_key(d, Q.curve)
@@ -119,10 +117,15 @@ class PEMEncoder(KeyEncoder):
         ec_private_key = asn1_structure(SEQUENCE, sequence)
         b64_data = "\n".join(wrap(b2a_base64(ec_private_key).decode(), 64))
 
-        return EC_PRIVATE_HEADER + "\n" + b64_data + "\n" + EC_PRIVATE_FOOTER
+        return (
+            self.EC_PRIVATE_HEADER
+            + b"\n"
+            + b64_data.encode()
+            + b"\n"
+            + self.EC_PRIVATE_FOOTER
+        )
 
-    @staticmethod
-    def decode_public_key(pemdata: str, curve: Optional[Curve] = None) -> Point:
+    def decode_public_key(self, pemdata: bytes, curve: Curve) -> Point:
         """Decode a PEM encoded public key as described in
         `RFC 5480 <https://tools.ietf.org/html/rfc5480>`_.
 
@@ -133,27 +136,30 @@ class PEMEncoder(KeyEncoder):
             (long, fastecdsa.point.Point): A private key, public key tuple. If the encoded key was a
             public key the first entry in the tuple is None.
         """
-        parsed = PEMEncoder._parse_ascii_armored_base64(pemdata)
-        PEMEncoder._parse_asn1_structure(parsed)
+        parsed = self._parse_ascii_armored_base64(pemdata)
+        self._parse_asn1_structure(parsed)
 
-        x, y, curve = None, None, None
-        for value_type, value in PEMEncoder.ASN1_PARSED_DATA:
-            if value_type == OBJECT_IDENTIFIER and curve is None:
-                curve = Curve.get_curve_by_oid(value)
+        x, y = None, None
+        for value_type, value in self.asn1_parsed_data:
+            if value_type == OBJECT_IDENTIFIER:
+                # override curve if explicitly defined in the encoded key
+                encoded_curve = Curve.get_curve_by_oid(value)
+                if encoded_curve is not None:
+                    encoded_curve = curve
+
             elif value_type == BIT_STRING:
                 value = value[2:]  # strip off b'\x00\x04'
-                x = int(hexlify(value[: len(value) // 2]), 16)
-                y = int(hexlify(value[len(value) // 2 :]), 16)
+                x = int.from_bytes(value[: len(value) // 2], "big")
+                y = int.from_bytes(value[len(value) // 2 :], "big")
 
-        PEMEncoder.ASN1_PARSED_DATA = []
+        self.asn1_parsed_data = []
 
         if curve is None or x is None or y is None:
             raise PEMEncoderError(f"Could not parse public key. {x=}, {y=}, {curve=}")
 
         return Point(x, y, curve)
 
-    @staticmethod
-    def decode_private_key(pemdata: str) -> Tuple[int, Optional[Point]]:
+    def decode_private_key(self, pemdata: bytes) -> int:
         """Decode a PEM encoded EC private key as described in
         `RFC 5915 <https://tools.ietf.org/html/rfc5915.html>`_.
 
@@ -161,30 +167,19 @@ class PEMEncoder(KeyEncoder):
             pemdata (bytes): A sequence of bytes representing an encoded EC key.
 
         Returns:
-            (long, fastecdsa.point.Point): A private key, public key tuple. If the encoded key was a
-            public key the first entry in the tuple is None.
+            int: The private key.
         """
-        parsed = PEMEncoder._parse_ascii_armored_base64(pemdata)
-        PEMEncoder._parse_asn1_structure(parsed)
+        parsed = self._parse_ascii_armored_base64(pemdata)
+        self._parse_asn1_structure(parsed)
 
-        d, Q, x, y, curve = None, None, None, None, None
-
-        for value_type, value in PEMEncoder.ASN1_PARSED_DATA:
+        d = None
+        for value_type, value in self.asn1_parsed_data:
             if value_type == OCTET_STRING:
                 d = int(hexlify(value), 16)
-            elif value_type == OBJECT_IDENTIFIER and curve is None:
-                curve = Curve.get_curve_by_oid(value)
-            elif value_type == BIT_STRING:
-                value = value[2:]  # strip off b'\x00\x04'
-                x = int(hexlify(value[: len(value) // 2]), 16)
-                y = int(hexlify(value[len(value) // 2 :]), 16)
 
-        PEMEncoder.ASN1_PARSED_DATA = []
+        self.asn1_parsed_data = []
 
         if d is None:
             raise PEMEncoderError("Could not parse private key.")
 
-        if x is not None and y is not None and curve is not None:
-            Q = Point(x, y, curve)
-
-        return d, Q
+        return d