+++ tests/test_prime_field_curve_math.py
@@ -2,8 +2,8 @@ from random import choice, randint
 from unittest import TestCase
 
 from . import CURVES
-from ..curve import P192, P224, P256, P384, P521, secp256k1, W25519, W448
-from ..point import Point
+from fastecdsa.curve import P192, P224, P256, P384, P521, secp256k1, W25519, W448
+from fastecdsa.point import Point
 
 
 class TestPrimeFieldCurve(TestCase):