+++ CHANGELOG.md
@@ -2,15 +2,26 @@
 
 ## [3.0.0]
 ### Added
-- `uv` for package management
 - Support for python3.13
+- `uv` for package management
 - Typing for all modules in this package
 - Github Actions for testing and publishing
 - Pre-commit hooks to enforce formatting and type checks
 
+### Changed
+- `fastecdsa.encodng.KeyEncoder` interface:
+  - All methods changed from static methods to instance methods
+  - Encoders always return `bytes`
+  - Decoders always take `bytes` for key data
+  - Private key methods do not consume or return public key data in their signatures
+    - Private key encoding _does_ take a curve parameter as some encoders do encode the public key as part of the encoded private key (in this case the public key is computed using the private key and the curve)
+  -
+
 ### Removed
 - Support for python3.7
 - Support for python3.8
+- `int` tuples are no longer allowed to be passed as `Q` in `ecdsa.verify`. `Q` must be of type `point.Point` now.
+- `keys.export_key` and `keys.import_key`. Superseded by functions specific to public and private keys.
 
 ## [2.3.2]
 ### Added